app/requirements.txt
soundcloud-dl
spotify-dl
python-dotenv